Refs #14951: Move payload Pool to subfolder
