+++ ddsrouter_core/src/cpp/efficiency/payload/MapPayloadPool.hpp
@@ -16,14 +16,14 @@
  * @file MapPayloadPool.hpp
  */
 
-#ifndef __SRC_DDSROUTERCORE_EFFICIENCY_MAPPAYLOADPOOL_HPP_
-#define __SRC_DDSROUTERCORE_EFFICIENCY_MAPPAYLOADPOOL_HPP_
+#ifndef __SRC_DDSROUTERCORE_EFFICIENCY_PAYLOAD_MAPPAYLOADPOOL_HPP_
+#define __SRC_DDSROUTERCORE_EFFICIENCY_PAYLOAD_MAPPAYLOADPOOL_HPP_
 
 #include <atomic>
 #include <map>
 #include <mutex>
 
-#include <efficiency/PayloadPool.hpp>
+#include <efficiency/payload/PayloadPool.hpp>
 
 namespace eprosima {
 namespace ddsrouter {
@@ -116,4 +116,4 @@ protected:
 } /* namespace ddsrouter */
 } /* namespace eprosima */
 
-#endif /* __SRC_DDSROUTERCORE_EFFICIENCY_MAPPAYLOADPOOL_HPP_ */
+#endif /* __SRC_DDSROUTERCORE_EFFICIENCY_PAYLOAD_MAPPAYLOADPOOL_HPP_ */
