+++ ddsrouter_core/src/cpp/efficiency/payload/PayloadPool.hpp
@@ -16,8 +16,8 @@
  * @file PayloadPool.hpp
  */
 
-#ifndef __SRC_DDSROUTERCORE_EFFICIENCY_PAYLOADPOOL_HPP_
-#define __SRC_DDSROUTERCORE_EFFICIENCY_PAYLOADPOOL_HPP_
+#ifndef __SRC_DDSROUTERCORE_EFFICIENCY_PAYLOAD_PAYLOADPOOL_HPP_
+#define __SRC_DDSROUTERCORE_EFFICIENCY_PAYLOAD_PAYLOADPOOL_HPP_
 
 #include <atomic>
 
@@ -245,4 +245,4 @@ protected:
 } /* namespace ddsrouter */
 } /* namespace eprosima */
 
-#endif /* __SRC_DDSROUTERCORE_EFFICIENCY_PAYLOADPOOL_HPP_ */
+#endif /* __SRC_DDSROUTERCORE_EFFICIENCY_PAYLOAD_PAYLOADPOOL_HPP_ */
