+++ ddsrouter_core/test/unittest/efficiency/FastPayloadPoolTest.cpp
@@ -17,8 +17,8 @@
 
 #include <fastdds/rtps/common/CacheChange.h>
 
-#include <efficiency/PayloadPool.hpp>
-#include <efficiency/FastPayloadPool.hpp>
+#include <efficiency/payload/PayloadPool.hpp>
+#include <efficiency/payload/FastPayloadPool.hpp>
 #include <ddsrouter_utils/exception/InconsistencyException.hpp>
 
 using namespace eprosima::ddsrouter;
